This PR improves k_quants perplexity scores by tweaking the quantization approach and quantization mixes. It is fully backward compatible (but obviously one needs to re-quantize the models to take advantage of these improvements).

The most significant gains are for `LLAMA_FTYPE_MOSTLY_Q2_K`, where perplexity is reduced by a significant margin while slightly reducing the model size (e.g., from 2.67 GiB to 2.63 GiB for 7B). See graphs below.

Significant improvements are also observed for `LLAMA_FTYPE_MOSTLY_Q3_K_M` and `LLAMA_FTYPE_MOSTLY_Q4_K_S` for LLaMA-v2-7B. This comes at the expense of a slightly increased model size (e.g., at 7B, 3.59 GiB vs 3.56 GiB for `Q4_K_S` and 3.07 GiB vs 3.06 GiB for `Q3_K_M`). 

Other quantization types / models are slightly better for LLaMA-v2 (but the change is much smaller compared to those mentioned above), or basically the same for LLaMA-v1.

Note on `LLAMA_FTYPE_MOSTLY_Q2_K`: strictly speaking, this is now mostly a `Q3_K` quantization. All tensors are quantized using `Q3_K`, except for attention `K` and `Q`, which are `Q2_K`, and `output.weight`, which is `Q6_K` as usual. I considered naming it `LLAMA_FTYPE_MOSTLY_Q3_K_XS` or similar, but given that this model is smaller and better than the previous `LLAMA_FTYPE_MOSTLY_Q2_K`, so the existing `Q2_K` model would have been useless in comparison, I decided that it is simpler to just re-use the `LLAMA_FTYPE_MOSTLY_Q2_K` designation for this new quantization mix.

The following graph shows perplexity vs model size for the LLaMA-v2-7B model and a context length of 512. Black dots/lines are for current master (i.e., after the merge of the `GGUF` related changes). Red dots/lines depict the results of this PR. Results for `Q4_0, Q4_1, Q5_0` and `Q5_1` on current master are shown in blue for comparison. The perplexity of the `fp16` model is 5.7963. The new `Q6_K` quantization arrives at 5.8067 (so, 0.18% higher) compared to 5.8118 (0.27% higher) on Master.
 
![ppl_vs_size_l2_new](https://github.com/ggerganov/llama.cpp/assets/48489457/8676b79c-48d0-48d9-a771-db14b8fc67b4)

The following graph is the same as the above, but with a smaller plot range to better appreciate the perplexity differences in the 4-6 bit quantization range.

![ppl_vs_size_l2_new1](https://github.com/ggerganov/llama.cpp/assets/48489457/a1df2343-d49f-4283-8ce6-4ba9f63d860e)

Similar to the above graphs, but for the LLaMA-v1-7B model.

![ppl_vs_size_l1_new](https://github.com/ggerganov/llama.cpp/assets/48489457/fa8ee614-c827-44ee-83a2-2fe2aa575b03)