@klosax 2023-08-22
> would it be possible to add some kind of meta information the the gguf to be able to determine if it was generated using the improvements of this pr

Currently main will print the number of tensors of each quantization format.