@ikawrakow 2023-08-22
> Currently main will print the number of tensors of each quantization format.

Yes, but one might decide to change the quantization strategy, so even though all tensors are quantized with the same type, result is still different. For instance, in this PR I have changed `Q4_K` and `Q5_K` to use the newly added function `make_qkx2_quants()` instead of the previous approach in `make_qkx1_quants()`. Hence, a date (or any other version identifier such as the commit hash) as requested by @Green-Sky will be a very useful thing to have in the meta data (and be printed, so people know what kind of quantized model they are using).

As it stands, when running with this PR, `LLAMA_FTYPE_MOSTLY_Q2_K` to `LLAMA_FTYPE_MOSTLY_Q3_K_L` get all reported as `Mostly Q3_K - Medium`, `LLAMA_FTYPE_MOSTLY_Q4_K_S` and `LLAMA_FTYPE_MOSTLY_Q4_K_M` are both reported as `Mostly Q4_K - Medium`, etc. This will be really confusing if it remains this way.