@ggerganov 2023-08-22
#2710  adds the `ftype` field back to the meta data.

Feel free to extend further the meta info with version/date/commit/etc. As long as the added KV info is optional, we can extend it anyway we like