@TheBloke 2023-08-22
Looks fatnastic! 6.0x for Q3 is amazing