@ikawrakow 2023-08-22
> Any benchmarks on models larger than 7B?

Yes, I did comparison for both 13B LLaMA's. Bud development was done on a branch that did not have the GGUF changes. When I was ready to submit the PR, I rebased on master, which brought in the GGUF changes, which changes the perplexity results. The change is actually quite dramatic for LLaMA-v2-13B: `fp16` perplexity on current master is `5.1195`, while what I had before was `PPL = 5.1000`. This is due to the change in `rms_eps`. Before GGUF, the default `rms_eps` was `5e-6`. After GGUF, it is taken from the model meta data and there is no way to modify it, so it ends up being `1e-5`. It was discussed e.g. in #2384 that `5e-6` gives lower perplexities than `1e-5`, despite `1e-5` having been used during training. In any case, I'm re-running all calculations for the 13B models and will post when they become ready.