@ikawrakow 2023-08-22
OK, here is the LLaMA-v2-13B result.

First an overview:
![ppl_vs_size_l2_13_new](https://github.com/ggerganov/llama.cpp/assets/48489457/930d273b-f78f-4b95-9524-2365dbf2eff7)

Then with focus on the 4-6 bit quantization range:
![ppl_vs_size_l2_13_new1](https://github.com/ggerganov/llama.cpp/assets/48489457/89546a4b-1011-4e4f-a06b-aa2099887596)