@IgnacioFDM 2023-08-22
Massive improvement with Q2_K it looks like