@Green-Sky 2023-08-22
the values in the help of the quantization tool where not updated. @ikawrakow