@mirek190 2023-08-22
q4.x or q5.x should be banned already as qk models are just better in everything ....