@Cebtenzzre 2023-08-23
I ran some performance tests. The most noticeable change is Q2_K, which is now 40% slower.
| GPU |   Model   |  Test | t/s before | t/s PR | Speedup   |
|:----|:----------|:------|:-----------|:-------|:----------|
| P40 | 7b q4_0   | tg128 | 55.94      | 55.91  | no change |
| P40 | 7b q2_K   | tg128 | 55.71      | 33.38  | 0.599     |
| P40 | 7b q3_K_S | tg128 | 30.84      | 30.81  | no change |
| P40 | 7b q3_K_M | tg128 | 37.35      | 37.20  | 0.996     |
| P40 | 7b q3_K_L | tg128 | 34.79      | 34.77  | no change |
| P40 | 7b q4_K_S | tg128 | 53.97      | 53.32  | 0.988     |
| P40 | 7b q4_K_M | tg128 | 49.12      | 49.13  | no change |
| P40 | 7b q5_K_S | tg128 | 42.90      | 42.90  | no change |
| P40 | 7b q5_K_M | tg128 | 41.06      | 41.07  | no change |
| P40 | 7b q6_K   | tg128 | 33.50      | 33.50  | no change |