@KerfuffleV2 2023-08-23
> The most noticeable change is Q2_K, which is now 40% slower.

That seems surprising since this is a backward compatible change. You should be able to quantize with this version and then test with a version before the pull was committed - if you do that, you still see a large performance difference?