@klosax 2023-08-23
I suggest all `+0.0004 ppl` should be the difference from F32 of the PTH model which gives the lowest possible ppl. The original PTH models are using BF16 which will be cut when converting to F16.