@ikawrakow 2023-08-23
A PPL difference of +/- 0.001 is within the noise for the amount of tokens in Wikitext. In the case of LLaMA-v1-7B it happens that `Q6_K` by chance arrives at a better PPL than `fp16` (and `Q8_0`). But this will to be the case in general. For instance, for LLaMA-v2-7B, `Q6_K` has PPL = 5.8067, which is 0.18% higher than the PPL = 5.7963 for `fp16` (and this is likely outside of the statistical uncertainty of the result, but one needs a proper uncertainty estimate added to the perplexity tool to confirm that). When editing the numbers I just took what was currently on master and adapted the number. Not sure why the previous explanations were removed. I was considering to update with the LLaMA-v2-7B numbers, but then decided to go for continuity and keep LLaMA-v1-7B results in the help.