@klosax 2023-08-23
The +/- ppl statistic may be is confusing for normal users to understand. Printing the real ppl may be better?